fix: move multi account features from settings into devtools
