+++ js/packages/i18n/locale/en-US/messages.json
@@ -176,9 +176,6 @@
 			"backup-button": "Backup",
 			"backup-notif-title": "Success",
 			"backup-notif-desc": "Account has been backed up!",
-			"accounts-button": "Accounts",
-			"create-button": "Create an account",
-			"import-button": "Import an account",
 			"link-button": "Link another device",
 			"delete-button": "Delete this account",
 			"delete-title": "Account deletion"
@@ -241,6 +238,9 @@
 			"header-left-button": "Device infos",
 			"header-middle-button": "Generate fake data",
 			"header-right-button": "Restart daemon",
+      "accounts-button": "Accounts",
+      "create-button": "Create an account",
+      "import-button": "Import an account",
 			"force-mock-button": "Force mock",
 			"simulate-button": "Simulate stream error",
 			"simulate-js-error-button": "Simulate JavaScript error",
